requirements.txt
snakemake==8.4.3
omni-schema==0.0.1
linkml-runtime>=1.7.0
rdflib==7.0.0

PyYAML~=6.0.1
networkx~=3.2.1
matplotlib~=3.8.3